fea_v5.0: 一体化雷达水位计设备: 添加功能页面

+++ app/src/main/java/com/shmedo/mcloudapp/ui/page/device/u_product/fragment/ud/UDSensorParamFragment.kt
@@ -217,7 +217,7 @@ class UDSensorParamFragment : BaseIOTDeviceFragment() {
             return
         }
         try {
-            val value = mStates.installAngleOffsetThreshold.get().toInt()
+            val value = mStates.installAngleOffsetThreshold.get().toDouble()
             if (value > 360) {
                 showMessageDialog("安装角度偏移阈值不能大于360!")
                 return
